Implemented directory Browser, more bug fixes

+++ Simple AD/TabContainers/ContainerUserReport.vb
@@ -1,13 +1,17 @@
 ﻿Public Class ContainerUserReport
     Inherits UserControl
 
-    Private _ID
-    Private _JobName
+    Private _ID As Integer
+    Private _Job As JobUserReport
+    Private _JobName As String
+    Private _Path As String
+
+    Private WithEvents _ControlDomainTreeContainer As ControlDomainTreeContainer
 
     Private DataTableSource As DataTable
 
-    Public Property ID
-        Set(value)
+    Public Property ID As Integer
+        Set(value As Integer)
             _ID = value
         End Set
         Get
@@ -15,12 +19,21 @@
         End Get
     End Property
 
-    Public Property JobName
-        Set(value)
+    Public Property JobName As String
+        Set(value As String)
             _JobName = value
         End Set
         Get
             Return _JobName
+        End Get
+    End Property
+
+    Public Property Path As String
+        Set(value As String)
+            _Path = value
+        End Set
+        Get
+            Return _Path
         End Get
     End Property
 
@@ -35,7 +48,7 @@
         End Get
     End Property
 
-    Public Sub New(ByVal ID As Integer, ByVal JobName As String)
+    Public Sub New(ByVal ID As Integer, ByVal JobName As String, ByVal Job As JobUserReport)
 
         Me.ID = ID
         Me.JobName = JobName
@@ -46,8 +59,10 @@
 
         MainDataGrid.DoubleBuffered(True)
 
-        Me.MainSplitContainer.Panel1.Controls.Add(New ControlDomainTreeContainer(Me))
-        Me.MainSplitContainer.Panel1Collapsed = True
+        _ControlDomainTreeContainer = New ControlDomainTreeContainer(Me)
+        _Job = Job
+
+        Me.DomainPl.Controls.Add(_ControlDomainTreeContainer)
     End Sub
 
     Public Function GetMainDataGrid() As DataGridView
@@ -81,55 +96,39 @@
     End Sub
 
     Private Sub MainDataGrid_CellFormatting(sender As Object, e As DataGridViewCellFormattingEventArgs) Handles MainDataGrid.CellFormatting
-        If MainDataGrid.Columns.Contains("Name") Then
-            If e.ColumnIndex = MainDataGrid.Columns("Name").Index Then
-                e.CellStyle.BackColor = Color.WhiteSmoke
-            End If
-        End If
-
-        If String.IsNullOrEmpty(MainDataGrid.Rows(e.RowIndex).Cells("sAMAccountName").Value.ToString) Then
-            MainDataGrid.Rows(e.RowIndex).DefaultCellStyle.ForeColor = SystemColors.ControlDarkDark
-        End If
-
         Try
-            If Not String.IsNullOrEmpty(MainDataGrid.Rows(e.RowIndex).Cells("sAMAccountType").Value.ToString) Then
+            If Not MainDataGrid.Rows(e.RowIndex).Cells("objectClass").Value Is Nothing Then
 
-                Select Case MainDataGrid.Rows(e.RowIndex).Cells("sAMAccountType").Value.ToString
-                    Case "805306368"
-                        If MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "546" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "514" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "66082" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "66050" Then
-                            With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                                .Image = IconDisabledUSer
-                            End With
-                        Else
-                            With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                                .Image = IconUser
-                            End With
-                        End If
-                    Case "805306369"
-                        With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                            .Image = IconComputer
-                        End With
-                    Case "268435456"
-                        With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                            .Image = IconGroup
-                        End With
-                    Case "268435457"
-                        With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                            .Image = IconGroup
-                        End With
-                    Case "536870912"
-                        With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                            .Image = IconGroup
-                        End With
-                    Case Else
-                        With DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
-                            .Image = IconContainer
-                        End With
-                End Select
+                Dim ObjectClass As String = MainDataGrid.Rows(e.RowIndex).Cells("objectClass").Value.ToString
+
+                If MainDataGrid.Columns.Contains("name") Then
+
+                    Dim CellImage As TextAndImageCell = DirectCast(MainDataGrid.Rows.Item(e.RowIndex).Cells("name"), TextAndImageCell)
+
+                    Select Case ObjectClass
+                        Case "user"
+                            If MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "546" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "514" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "66082" Or MainDataGrid.Rows(e.RowIndex).Cells("UserAccountControl").Value = "66050" Then
+                                CellImage.Image = IconDisabledUSer
+                            Else
+                                CellImage.Image = IconUser
+                            End If
+                        Case "computer"
+                            CellImage.Image = IconComputer
+                        Case "group"
+                            CellImage.Image = IconGroup
+                        Case "container"
+                            CellImage.Image = IconContainer
+                        Case "organizationalUnit"
+                            CellImage.Image = IconOU
+                        Case "contact"
+                            CellImage.Image = IconContact
+                        Case Else
+                            CellImage.Style.ForeColor = SystemColors.ControlDarkDark
+                    End Select
+                End If
             End If
-
         Catch Ex As Exception
-            Debug.WriteLine("[Error] " & Ex.Message)
+            Debug.WriteLine("[Error] CellFormatting Failed: " & Ex.Message)
         End Try
     End Sub
 
@@ -137,7 +136,7 @@
         Try
             If Not IsDBNull(MainDataGrid.SelectedRows(0).Cells("sAMAccountName").Value) Then
                 Dim Sam = MainDataGrid.SelectedRows(0).Cells("sAMAccountName").Value
-                Dim Name = MainDataGrid.SelectedRows(0).Cells("Name").Value
+                Dim Name = MainDataGrid.SelectedRows(0).Cells("name").Value
                 Dim ShowUserProps = New FormUserAttributes(Sam, Name, MainDataGrid.SelectedRows(0).Index)
             End If
         Catch Ex As Exception
@@ -197,8 +196,13 @@
     End Function
 
     Private Sub ToolStripMenuItem4_Click(sender As Object, e As EventArgs) Handles ToolStripMenuItem4.Click
-        If DeleteADObject(GetSelectedUser()) Then
-            MainDataGrid.Rows.RemoveAt(MainDataGrid.SelectedRows(0).Index)
+        Dim User = GetSelectedUser()
+        Dim DeleteForm = New FormConfirmation("Are you sure you wish to delete " & User & "?", ConfirmationType.delete)
+        DeleteForm.ShowDialog()
+        If DeleteForm.DialogResult = DialogResult.Yes Then
+            If DeleteADObject(User) Then
+                MainDataGrid.Rows.RemoveAt(MainDataGrid.SelectedRows(0).Index)
+            End If
         End If
     End Sub
 
@@ -209,9 +213,7 @@
 
             For Each Row As DataGridViewRow In MainDataGrid.SelectedRows
                 If Not String.IsNullOrEmpty(Row.Cells("sAMAccountName").Value) Then
-                    Dim User = New UserRow
-                    User.Username = Row.Cells("sAMAccountName").Value
-                    User.Row = Row
+                    Dim User = New UserRow With {.Username = Row.Cells("sAMAccountName").Value, .Row = Row}
                     UserArray.Add(User)
                 End If
             Next
@@ -249,12 +251,38 @@
     End Sub
 
     Private Sub DeleteToolStripMenuItem_Click(sender As Object, e As EventArgs) Handles DeleteToolStripMenuItem.Click
-        For Each User As UserRow In GetSelectedUsers()
-            If DeleteADObject(User.Username) Then
-                MainDataGrid.Rows.Remove(User.Row)
-            End If
-        Next
+        Dim SelectedUsers As List(Of UserRow) = GetSelectedUsers()
+        Dim DeleteForm = New FormConfirmation("Are you sure you wish to delete " & SelectedUsers.Count & " objects?", ConfirmationType.Delete)
+        DeleteForm.ShowDialog()
+        If DeleteForm.DialogResult = DialogResult.Yes Then
+            For Each User As UserRow In SelectedUsers
+                If DeleteADObject(User.Username) Then
+                    MainDataGrid.Rows.Remove(User.Row)
+                End If
+            Next
+        End If
     End Sub
+
+    Private Sub MoveSingleToolStripMenuItem_Click(sender As Object, e As EventArgs) Handles MoveSingleToolStripMenuItem.Click
+        Dim User = GetSelectedUser()
+        Dim MoveForm = New FormMoveObject
+        MoveForm.ShowDialog()
+        If MoveForm.DialogResult = DialogResult.Yes Then
+            If MoveADObject(User, MoveForm.SelecetdOU) = True Then
+                Dim ResultForm = New FormAlert(User & " has been moved to:" & Environment.NewLine & MoveForm.SelecetdOU, AlertType.Success)
+                ResultForm.ShowDialog()
+            Else
+                Dim ResultForm = New FormAlert("An Error occured while trying to move" & User & " to:" & Environment.NewLine & MoveForm.SelecetdOU, AlertType.ErrorAlert)
+                ResultForm.ShowDialog()
+            End If
+        End If
+    End Sub
+
+    Public Sub SelecetdOu_changed(ByVal Path As String) Handles _ControlDomainTreeContainer.SelectedOUChanged
+        Me.Path = Path
+        _Job.Refresh(Path)
+    End Sub
+
 End Class
 
 Public Class UserRow
